Server: boolean parameter;services:bool->flag;checker:flag
Previously it was a bit of a mess, but now the server passes a boolean
to services. False means 'provide just displayString', true
means 'provide displayString and simpleDisplayString'.

The services layer passes the boolean around, plus uses it to decide
whether to call the compiler in simple mode.

The compiler uses TypeFormatFlags.SuperSimple exclusively to decide whether to
simplify. But it never introduces SuperSimple; the services needs to
provide every time it calls into the compiler.

Note: It's now apparent that the compiler is a bit wrong. For example,
it converts almost every *symbol* to object, which is not correct.